separate test root for modules
