+++ jest.config.ts
@@ -4,7 +4,12 @@ const config: Config.InitialOptions = {
   preset: "ts-jest/presets/default-esm",
   setupFilesAfterEnv: ["<rootDir>/tests/setupTests.js"],
   testEnvironment: "node",
-  testPathIgnorePatterns: ["/node_modules/", "/dist/", "/api/"],
+  testPathIgnorePatterns: [
+    "/node_modules/",
+    "/dist/",
+    "/api/",
+    "/tests-modules/",
+  ],
   extensionsToTreatAsEsm: [".ts"],
   globals: {
     "ts-jest": {
