+++ package.json
@@ -47,8 +47,8 @@
     "prepublishOnly": "yarn build && yarn generateSchema",
     "test": "node --experimental-vm-modules node_modules/.bin/jest",
     "test:ts": "tsc --noEmit",
-    "test:esm": "node --experimental-vm-modules node_modules/.bin/jest tests/modules/esm/tests/*",
-    "test:cjs": "node node_modules/.bin/jest -c tests/modules/cjs/jest.config.js tests/modules/cjs/tests/*",
+    "test:esm": "node --experimental-vm-modules node_modules/.bin/jest -c tests-modules/esm/jest.config.js tests-modules/esm/tests/*",
+    "test:cjs": "node node_modules/.bin/jest -c tests-modules/cjs/jest.config.js tests-modules/cjs/tests/*",
     "test:modules": "yarn test:esm && yarn test:cjs",
     "test:build": "yarn test:modules"
   },
