+++ tests-modules/cjs/package.json
