+++ tests-modules/esm/package.json
