+++ tests-modules/esm/tests/tests.spec.js
@@ -1,4 +1,4 @@
-import yahooFinance from "../../../../dist/esm/src/index-node.js";
+import yahooFinance from "../../../dist/esm/src/index-node.js";
 
 describe("esm", () => {
   it("sanity", () => {
